0.1.7 Added async webserver and littlefs

+++ include/information/krInfo.h
@@ -9,8 +9,8 @@ class Information {
 
         struct System
         {
-            int uptimeSeconds;
-            long wifiRSSI;
+            uint32_t uptimeSeconds;
+            int8_t wifiRSSI;
             String IPAddress;
         };
         struct AudioPlayer
